Made the server welcome page template and refactor the server code
