+++ .npmignore
@@ -1,0 +1,2 @@
+icon.png
+teleport-flask-websocket.png
