+++ backend/servers/flask-websocket/config/_p_teleport.json
@@ -1,0 +1,3 @@
+{
+  "templates": $[project.templates]
+}
