+++ backend/servers/flask-websocket/scripts/_type_start.sh
