+++ package.json
@@ -7,5 +7,5 @@
     "type": "git",
     "url": "https://github.com/snipsco/teleport-flask-websocket"
   },
-	"version": "0.1.14"
+	"version": "0.1.17"
 }
